requirements.txt
django>=4.2.0
channels==4.0.0
channels-redis>=4.1.0
redis>=4.6.0
mysqlclient>=2.2.0
uvicorn[standard]>=0.25.0
daphne>=4.0.0 